cli: Extract Redis schema versioning
